refactor: Renamed total-fees to limit-trade-details
