--- packages/web/components/place-limit-tool/total-fees.tsx
@@ -1,105 +1,0 @@
-import { Dec, PricePretty } from "@keplr-wallet/unit";
-import { DEFAULT_VS_CURRENCY } from "@osmosis-labs/server";
-import classNames from "classnames";
-import React, { FC, memo, useMemo, useState } from "react";
-
-import { Icon } from "~/components/assets";
-import { SkeletonLoader } from "~/components/loaders/skeleton-loader";
-import { useTranslation } from "~/hooks";
-import { PlaceLimitState } from "~/hooks/limit-orders";
-import { formatPretty } from "~/utils/formatter";
-
-interface TotalFeesProps {
-  swapState: PlaceLimitState;
-}
-
-export const TotalFees: FC<TotalFeesProps> = memo(({ swapState }) => {
-  const { t } = useTranslation();
-  const [displayInfo, setDisplayInfo] = useState<boolean>(false);
-  const { makerFeeFiat, totalFeeFiat } = useMemo(() => {
-    const makerFeeFiat = (
-      swapState.paymentFiatValue ??
-      new PricePretty(DEFAULT_VS_CURRENCY, new Dec(0))
-    ).mul(swapState.makerFee);
-
-    const totalFeeFiat = makerFeeFiat;
-    return {
-      makerFeeFiat,
-      totalFeeFiat,
-    };
-  }, [swapState.makerFee, swapState.paymentFiatValue]);
-
-  const isLoading = useMemo(() => {
-    return swapState.isMakerFeeLoading;
-  }, [swapState.isMakerFeeLoading]);
-
-  return (
-    <SkeletonLoader isLoaded={!isLoading}>
-      <div className="flex w-full flex-col">
-        <div
-          className="flex w-full cursor-pointer items-center justify-between"
-          onClick={() => setDisplayInfo(!displayInfo)}
-        >
-          <div className="text-subtitle1">{t("place-limit.tradeDetails")}</div>
-          <div className="flex items-center justify-end text-body2 text-osmoverse-300">
-            <Icon
-              id="chevron-down"
-              className={classNames(
-                "ml-3 text-osmoverse-400 transition-transform",
-                {
-                  "rotate-180": displayInfo,
-                }
-              )}
-              height={12}
-              width={12}
-            />
-          </div>
-        </div>
-        <div
-          className={classNames(
-            "flex w-full flex-col items-start overflow-hidden transition-all",
-            {
-              "h-0": !displayInfo,
-              "h-[105px]": displayInfo,
-            }
-          )}
-        >
-          <div className="flex w-full items-center justify-between py-3 text-body2 text-osmoverse-300">
-            <div>
-              Total fees when filled ( ~
-              {formatPretty(swapState.makerFee.mul(new Dec(100)), {
-                maxDecimals: 2,
-                minimumFractionDigits: 2,
-              })}
-              %)
-            </div>
-            <div>
-              ~
-              {formatPretty(makerFeeFiat, {
-                maxDecimals: 2,
-                minimumFractionDigits: 2,
-              })}
-            </div>
-          </div>
-          <hr className="my-2 w-full text-osmoverse-700" />
-          <div className="flex w-full items-center justify-between py-3 text-body2 text-osmoverse-300">
-            <div>Receive</div>
-            <div>
-              ~
-              {formatPretty(swapState.expectedFiatAmountOut, {
-                maxDecimals: 2,
-                minimumFractionDigits: 2,
-              })}{" "}
-              (
-              {formatPretty(swapState.expectedTokenAmountOut, {
-                maxDecimals: 2,
-                minimumFractionDigits: 2,
-              })}
-              )
-            </div>
-          </div>
-        </div>
-      </div>
-    </SkeletonLoader>
-  );
-});
